delete add and updated complete

+++ src/Pages/MyCart.jsx
@@ -17,24 +17,36 @@ const MyCart = () => {
   }, [user, control]);
 
   const handleDelete = (id) => {
-    fetch(`http://localhost:5000/delete/${id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        if (data.deletedCount > 0) {
-          setControl(!control);
-          Swal.fire({
-            title: "Deleted!",
-            text: "Your Tourist Spot has been deleted.",
-            icon: "success",
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`http://localhost:5000/delete/${id}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            if (data.deletedCount > 0) {
+              setControl(!control);
+              Swal.fire({
+                title: "Deleted!",
+                text: "Your Tourist Spot has been deleted.",
+                icon: "success",
+              });
+            }
           });
-        }
+      }
 
-        // // remove
-        // const remainingData = users.filter((item) => item._id !== id);
-        // setUsers(remainingData);
-      });
+      // // remove
+      // const remainingData = users.filter((item) => item._id !== id);
+      // setUsers(remainingData);
+    });
   };
 
   return (
